Credential Issuance API and MVP Architecture Guide/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
qrcode[pil]==7.4.2
redis==5.0.1
celery==5.3.4
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
jinja2==3.1.2
reportlab==4.0.7
weasyprint==60.2
jsonschema==4.20.0

